+++ Soul Cards/Cartas/src/main/java/com/example/cartas/Inimigo.java
@@ -1,6 +1,6 @@
 package com.example.cartas;
 
 public class Inimigo {
-    int hpEnemy = 20;
+    double hpEnemy = 20;
     int energiaEnemy = 1;
 }
